Revert "Custom hook vs context api (#1)"
This reverts commit 95c7980c69f3a22ce5fda3cec3af9478625590b9.
