+++ src/client/components/start-context/app-context.js
@@ -1,0 +1,5 @@
+import React from "react"
+
+const appContext = React.createContext()
+
+export default appContext
