+++ src/client/components/start-context/current-user-nav.js
@@ -1,20 +1,21 @@
 import React from "react"
-import {useAppContext} from "./use-context-hook"
+import AppContext from "./app-context"
 
-const CurrentUserNav = () => {
-  let {user, busy, refresh} = useAppContext()
-  return (
-    user ? (
-      <span className="nav-text mr-3">Hi {user?.username}!</span>
-    ) : (
-      <span className="nav-text mr-3">
-        {busy ? (
-          <i className="fa fa-spin fa-sync" />
-        ) : (
-          <a href="#" onClick={(e) => {e.preventDefault(); refresh()}}>Login</a>
-        )}
-      </span>)
-  )
-}
+const CurrentUserNav = () => (
+  <AppContext.Consumer>
+    {({context, refresh}) => (
+      context.user ? (
+        <span className="nav-text mr-3">Hi {context?.user?.username}!</span>
+      ) : (
+        <span className="nav-text mr-3">
+          {context.busy ? (
+            <i className="fa fa-spin fa-sync" />
+          ) : (
+            <a href="#" onClick={(e) => {e.preventDefault(); refresh()}}>Login</a>
+          )}
+        </span>)
+    )}
+  </AppContext.Consumer>
+)
 
 export default CurrentUserNav
